L01 latest rev - issues with while loop
