+++ students/philip_behrend/lesson01/inventory_management/inventory_class.py
@@ -1,0 +1,19 @@
+# Inventory class
+class Inventory:
+    """Inventory class"""
+    def __init__(self, product_code, description, market_price, rental_price):
+        """init docstring"""
+        self.product_code = product_code
+        self.description = description
+        self.market_price = market_price
+        self.rental_price = rental_price
+
+    def return_as_dictionary(self):
+        """return inputs as dict"""
+        output_dict = {}
+        output_dict['product_code'] = self.product_code
+        output_dict['description'] = self.description
+        output_dict['market_price'] = self.market_price
+        output_dict['rental_price'] = self.rental_price
+
+        return output_dict
