--- students/philip_behrend/lesson01/inventory_management/temperature_sensor.py
@@ -1,2 +1,0 @@
-def read_temperature():
-    return 0
